Expose websockets to partial class
