+++ Equality/Services/ColumnService.cs
@@ -6,11 +6,6 @@ namespace Equality.Services
 {
     internal class ColumnService : ColumnServiceBase<Column, Board>, IColumnService
     {
-        public ColumnService(IApiClient apiClient, ITokenResolver tokenResolver)
-            : base(apiClient, tokenResolver)
-        {
-        }
-
         public ColumnService(IApiClient apiClient, ITokenResolver tokenResolver, IWebsocketClient websocketClient)
             : base(apiClient, tokenResolver, websocketClient)
         {
