+++ Equality/Services/Interfaces/IColumnService.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-using Equality.Models;
+﻿using Equality.Models;
 
 namespace Equality.Services
 {
